v1/requirements.txt
openai>=1.0.0
python-docx>=0.8.11
markdown2>=2.4.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1
Jinja2>=3.1.2